Major overhaul Added a resource_map that holds links to devices and such. Made "initialize" a pure virtual function in the Component class. Simplified the code for adding objects. Fixed volume adjustments and made xml settings. Added some sound effects. Still working on a few more.

+++ Source/SoundDevice.h
@@ -20,8 +20,8 @@ public:
 
 	void setAsBackground(std::string &background);
 	void setAsBackground(BackgroundMusic* background);
-	Uint8 setMusicVolume(Uint8 volume);
-	Uint8 setSoundVolume(Uint8 volume);
+	int setMusicVolume(int volume);
+	int setSoundVolume(int volume);
 
 	bool removeSound(std::string &name);
 	bool initialized{ false };
